Refactored model and standardmodel
